renamed re and restraints CLIs for documentation purposes
