+++ src/haddock/clis/re/score.py
